Move exceptions into `API\Model\Exceptions`
