+++ includes/api/model/exceptions/class-bh-wp-bitcoin-gateway-exception.php
@@ -7,7 +7,7 @@
  * @package brianhenryie/bh-wp-bitcoin-gateway
  */
 
-namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model;
+namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Exceptions;
 
 use Exception;
