+++ includes/api/model/exceptions/class-rate-limit-exception.php
@@ -9,7 +9,7 @@
  * @package brianhenryie/bh-wp-bitcoin-gateway
  */
 
-namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model;
+namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Exceptions;
 
 use DateTimeInterface;
 use Throwable;
